hw-5

+++ api/endpoints/academic_performance.py
@@ -1,4 +1,8 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Security
+from fastapi.exceptions import HTTPException
+from fastapi_jwt import JwtAuthorizationCredentials
+
+from api.jwt import access_security, is_token_blacklisted
 
 from schemas import AcademicPerformance
 from services.academic_performance_service import AcademicPerformanceService
@@ -10,7 +14,13 @@ academic_performance_router.tags = ["AcademicPerformance"]
 
 
 @academic_performance_router.delete("/", response_model=AcademicPerformance)
-async def get_academic_performance(academic_performance_id: int):
+async def get_academic_performance(
+        academic_performance_id: int,
+        credentials: JwtAuthorizationCredentials = Security(access_security)
+):
+    if is_token_blacklisted(token=credentials.jti):
+        raise HTTPException(status_code=400, detail="User not authorized")
+
     academic_performance = await academic_performance_service.get_academic_performance(
         academic_performance_id=academic_performance_id
     )
@@ -18,7 +28,13 @@ async def get_academic_performance(academic_performance_id: int):
 
 
 @academic_performance_router.delete("/", response_model=AcademicPerformance)
-async def delete_academic_performance(academic_performance_id: int):
+async def delete_academic_performance(
+        academic_performance_id: int,
+        credentials: JwtAuthorizationCredentials = Security(access_security)
+):
+    if is_token_blacklisted(token=credentials.jti):
+        raise HTTPException(status_code=400, detail="User not authorized")
+
     academic_performance = await academic_performance_service.delete_academic_performance(
         academic_performance_id=academic_performance_id
     )
@@ -26,7 +42,13 @@ async def delete_academic_performance(academic_performance_id: int):
 
 
 @academic_performance_router.put("/", response_model=AcademicPerformance)
-async def add_academic_performance(academic_performance: AcademicPerformance):
+async def add_academic_performance(
+        academic_performance: AcademicPerformance,
+        credentials: JwtAuthorizationCredentials = Security(access_security)
+):
+    if is_token_blacklisted(token=credentials.jti):
+        raise HTTPException(status_code=400, detail="User not authorized")
+
     academic_performance = await academic_performance_service.add_academic_performance(
         academic_performance=academic_performance
     )
@@ -34,7 +56,13 @@ async def add_academic_performance(academic_performance: AcademicPerformance):
 
 
 @academic_performance_router.patch("/", response_model=AcademicPerformance)
-async def update_academic_performance(academic_performance: AcademicPerformance):
+async def update_academic_performance(
+        academic_performance: AcademicPerformance,
+        credentials: JwtAuthorizationCredentials = Security(access_security)
+):
+    if is_token_blacklisted(token=credentials.jti):
+        raise HTTPException(status_code=400, detail="User not authorized")
+
     academic_performance = await academic_performance_service.update_academic_performance(
         academic_performance=academic_performance
     )
